Fix bug in ActivityDeleteCommand
ActivityDeleteCommand was not removing ActivityCosts from the payer if
the payer was not part of the list of participants, resulting in
obsolete ActivityCosts remaining in the payer's Person object.
This fixes that by removing the appropriate ActivityCosts from the payer
if the payer is not included in the list of participants.