Change Search JUnit tests
Previously, SearchCommand used its output for JUnit tests.

It is not very robust.

A List<Item> results is implemented instead to represent the returned items for commands that do not modify ItemList. Examples are list and search.

More tests could be done to verify the printing functionality. However, the new implementation suffices in checking that the order of selected Items are correct.

Let's
- Create a List<Item> results in the base Command.class
- Change SearchCommand tests to check against this results list.