Implement IdGenerator and changing add appt cmd
IdGenerator has a static method to generator a pseudo unique ID which
is a 8 digit number in String format. Can be used for other lists as
well. Still in the midst of changing the add appt cmd. Currently, it
should check if patient and doctor exists. However, still need to
implement adding into the appointment list.