Update generateItemList for use in stubs
Java, by default, adds references of objects into arraylist. This means the original Items in ItemList may be modified.

This may result in a race condition! Each test are given their own threads to execute in JUnit.

To fix this, we should clone the Item and add them to the list before returning the itemlist.

Let's
- Implement clone() for BorrowRecord, needed for deep copying
- Implement a deep copy of Item() in its clone method to clone BorrowRecord
- Change ItemListStubs to generate new mutable ItemLists that are clones of the specified Items.