Update ListFutureBorrowingsCommand.java
Resolves #151

Programme was initially unable to detect when there are no relevant
records in the future (and for the particular borrower). This does not
align with the User Guide where we should be able to.

Let's
* Use Java Streams to create a compiled list of relevant borrow records,
and use that list to check if we should return a message saying that
there are no records.